refactor: Subject in Teacher Subject Item
